+++ app/code/Deity/MagentoApi/Api/Sales/OrderManagementInterface.php
@@ -1,24 +1,19 @@
 <?php
+declare(strict_types=1);
 
 namespace Deity\MagentoApi\Api\Sales;
 
+/**
+ * Interface OrderManagement
+ * @package Deity\MagentoApi\Api\Sales
+ */
 interface OrderManagementInterface
 {
     /**
-     * @param mixed $orderId
-     * @return \Magento\Sales\Api\Data\OrderInterface
-     */
-    public function getItem($orderId);
-
-    /**
-     * @param \Magento\Framework\Api\SearchCriteria $searchCriteria
-     * @return \Magento\Sales\Api\Data\OrderSearchResultInterface
-     */
-    public function getCustomerOrders(\Magento\Framework\Api\SearchCriteria $searchCriteria);
-
-    /**
+     * Get order_id from paypal hash
+     *
      * @param string $paypalHash
      * @return int
      */
-    public function getOrderIdFromHash($paypalHash);
+    public function getOrderIdFromHash(string $paypalHash): int;
 }
