+++ app/code/Deity/MagentoApi/etc/webapi.xml
@@ -29,24 +29,10 @@
         </resources>
     </route>
 
-    <route url="/V1/orders/:orderId/order-info" method="GET">
-        <service class="\Deity\MagentoApi\Api\Sales\OrderManagementInterface" method="getItem"/>
-        <resources>
-            <resource ref="self" />
-        </resources>
-    </route>
-
     <route url="/V1/orders/get-order-from-paypal-hash/:paypalHash" method="GET">
         <service class="\Deity\MagentoApi\Api\Sales\OrderManagementInterface" method="getOrderIdFromHash"/>
         <resources>
             <resource ref="anonymous" />
-        </resources>
-    </route>
-
-    <route url="/V1/orders/mine" method="GET">
-        <service class="\Deity\MagentoApi\Api\Sales\OrderManagementInterface" method="getCustomerOrders"/>
-        <resources>
-            <resource ref="self" />
         </resources>
     </route>
 </routes>
