+++ app/code/Deity/QuoteApi/Test/Api/CartManagementTest.php
@@ -23,7 +23,7 @@ class CartManagementTest extends WebapiAbstract
     /**
      * @var ObjectManagerInterface
      */
-    protected $objectManager;
+    private $objectManager;
 
     /**
      *  setup before every test run. Update app config
