+++ app/code/Deity/Sales/etc/di.xml
@@ -2,6 +2,7 @@
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <preference for="Deity\SalesApi\Api\Data\OrderIdMaskInterface" type="Deity\Sales\Model\OrderIdMask" />
     <preference for="Deity\SalesApi\Api\OrderIdMaskRepositoryInterface" type="Deity\Sales\Model\OrderIdMaskRepository" />
+    <preference for="Deity\SalesApi\Api\OrderManagementInterface" type="Deity\Sales\Model\OrderManagement" />
 
     <type name="Magento\Sales\Api\OrderManagementInterface">
         <plugin name="deity_magento_order_management" type="Deity\Sales\Plugin\Api\OrderManagement" />
