+++ app/code/Deity/SalesApi/etc/webapi.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<routes xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Webapi:etc/webapi.xsd">
+    <route url="/V1/orders/mine" method="GET">
+        <service class="\Deity\SalesApi\Api\OrderManagementInterface" method="getCustomerOrders"/>
+        <resources>
+            <resource ref="self" />
+        </resources>
+    </route>
+    <route url="/V1/orders/:orderId/order-info" method="GET">
+        <service class="\Deity\SalesApi\Api\OrderManagementInterface" method="getItem"/>
+        <resources>
+            <resource ref="self" />
+        </resources>
+    </route>
+</routes>
